ci: simplify pass on the haul build-lane tooling
- components.txt: drop the phantom xnat-plugin-installer (removed from the
  image matrix by the XNAT 3.0.0 adoption) and correct the count to 8, so the
  scaffolding does not ship pre-drifted from the &image-matrix.
- move ghcr_digest from resolve.py into build_haul.py so resolve.py is the pure,
  I/O-free assembler its docstring describes and the live-inspect carry source
  sits in the glue layer; document the unused token param's intent.
- parse_fresh: drop the dead `or line` fallback (the docker-push format always
  emits `<name> <ref>`).

Deliberately skipped (agent consensus): sharing the empty/duplicate guard across
the two intentionally-independent stdlib modules, the fail-closed pin checks
(kept per-module by design), and lookup concurrency (not a hot path). 30 tests.
